Feat: Split the app into components
- Created three components for header, resultsTable and userInput.
